refactoring logging
