+++ frontend/src/api/FetchProfile.tsx
@@ -8,12 +8,17 @@ export interface UserProfile {
 const fetchProfile = async () => {
   try {
     const token = localStorage.getItem("token");
-    const response = await axios.get<UserProfile>(RestUrl.getProfile, {
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
-    });
-    return response.data;
+    if (token) {
+      const response = await axios.get<UserProfile>(RestUrl.getProfile, {
+        headers: {
+          Authorization: `Bearer ${token}`,
+        },
+      });
+      return response.data;
+    } else {
+      console.log("Empty token");
+      return null;
+    }
   } catch (error) {
     console.error("Error fetching profile:", error);
     return null;
