+++ frontend/src/api/RestUrl.tsx
@@ -2,6 +2,6 @@ const backendUrl = "http://localhost:3000";
 
 export const RestUrl = {
   postLogin: `${backendUrl}/auth/login`,
-  getProfile: `${backendUrl}auth/profile`,
+  getProfile: `${backendUrl}/auth/profile`,
   getValidCourses: `${backendUrl}/courses/valid`,
 };
