+++ frontend/src/page/main/MainPage.tsx
@@ -1,5 +1,5 @@
-import Footer from "../Footer";
-import Header from "../Header";
+import Footer from "../components/Footer";
+import Header from "../components/Header";
 import Cards from "./components/Cards";
 import BrandDescription from "./components/BrandDescription";
 import "./mainpage.css";
